* master: (233 commits)
  chore(deps): update dependency @types/ws to v7.4.4
  fix(deps): update engines to v2.23.0-27.ed792f504cb865b38d9d632b33acb6a65cedfc8c
  chore(migrate): change wording for unexecutable change (#7023)
  fix(deps): update engines to v2.23.0-25.1543ce559195fa2ed80b14020bddfa2675a34981
  fix(deps): update engines to v2.23.0-22.092cec160f64118d8fc06496ee4a9e8c25eca28a (#6971)
  chore(deps): update dependency @prisma/studio-server to v0.387.0 (#7024)
  chore: update pnpm-lock [skip ci]
  chore(deps): update dependency @prisma/studio-server to v0.386.0
  chore(deps): update dependency @prisma/studio-server to v0.384.0 (#7013)
  ci: gh ignore template changes
  chore(deps): update dependency lint-staged to v11 (#7006)
  chore(deps): update dependency lint-staged to v11 (#6987)
  chore(deps): update react-prisma
  chore(deps): update dependency @prisma/studio-server to v0.383.0
  chore(deps): update dependency @prisma/studio-server to v0.382.0 (#6990)
  fix(cli,migrate): regression on sqlserver datasource parsing (#6952)
  chore(deps): update dependency @prisma/studio-server to v0.380.0
  Update bug_report.yml
  feat: use new github issue forms (#6973)
  fix(deps): update dependency dotenv to v9 (#6955)
  ...

+++ src/packages/client/src/generation/generateClient.ts
@@ -15,7 +15,6 @@ import { promisify } from 'util'
 import { DMMF as PrismaClientDMMF } from '../runtime/dmmf-types'
 import { Dictionary } from '../runtime/utils/common'
 import { resolveDatasources } from '../utils/resolveDatasources'
-import { extractSqliteSources } from './extractSqliteSources'
 import { getPrismaClientDMMF } from './getDMMF'
 import { JS, TS, TSClient } from './TSClient'
 import { BrowserJS } from './TSClient/Generatable'
@@ -76,17 +75,14 @@ export async function buildClient({
   activeProvider,
 }: GenerateClientOptions): Promise<BuildClientResult> {
   const document = getPrismaClientDMMF(dmmf)
-  const useNapi = generator?.previewFeatures?.includes('napi')
+  const useNapi =
+    generator?.previewFeatures?.includes('nApi') ||
+    process.env.PRISMA_FORCE_NAPI === 'true'
   const client = new TSClient({
     document,
     runtimePath,
     browser,
     datasources: resolveDatasources(datasources, schemaDir, outputDir),
-    sqliteDatasourceOverrides: extractSqliteSources(
-      datamodel,
-      schemaDir,
-      outputDir,
-    ),
     generator,
     platforms: useNapi
       ? Object.keys(binaryPaths.libqueryEngineNapi!)
@@ -154,7 +150,9 @@ export async function generateClient({
   activeProvider,
 }: GenerateClientOptions): Promise<BuildClientResult | undefined> {
   const useDotPrisma = testMode ? !runtimePath : !generator?.isCustomOutput
-  const useNAPI = generator?.previewFeatures?.includes('napi')
+  const useNAPI =
+    generator?.previewFeatures?.includes('nApi') ||
+    process.env.PRISMA_FORCE_NAPI === 'true'
   runtimePath =
     runtimePath || (useDotPrisma ? '@prisma/client/runtime' : './runtime')
 
@@ -247,9 +245,16 @@ export async function generateClient({
     )
   }
   if (transpile) {
-    for (const filePath of Object.values(enginePath)) {
+    if (process.env.NETLIFY) {
+      await makeDir('/tmp/prisma-engines')
+    }
+
+    for (const [binaryTarget, filePath] of Object.entries(enginePath)) {
       const fileName = path.basename(filePath)
-      const target = path.join(finalOutputDir, fileName)
+      const target =
+        process.env.NETLIFY && binaryTarget !== 'rhel-openssl-1.0.x'
+          ? path.join('/tmp/prisma-engines', fileName)
+          : path.join(finalOutputDir, fileName)
       const [sourceFileSize, targetFileSize] = await Promise.all([
         fileSize(filePath),
         fileSize(target),
@@ -257,8 +262,12 @@ export async function generateClient({
 
       // If the target doesn't exist yet, copy it
       if (!targetFileSize) {
-        await copyFile(filePath, target)
-        continue
+        if (fs.existsSync(filePath)) {
+          await copyFile(filePath, target)
+          continue
+        } else {
+          throw new Error(`File at ${filePath} is required but was not present`)
+        }
       }
 
       // If target !== source size, they're definitely different, copy it
